Move functions to convex adam utils script
